e-boda-backend/backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
mongoengine==0.27.0
pymongo==4.6.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Environment variables
python-dotenv==1.0.0

# WebSocket support
websockets==12.0

# CORS
fastapi-cors==0.0.6

# Logging
python-json-logger==2.0.7

# Validation
pydantic==2.7.1
email-validator==2.3.0

# HTTP client (for future Google Maps integration)
httpx==0.26.0
